remove deprecated rest stuff
