+++ api/stream.go
@@ -1,0 +1,40 @@
+package api
+
+import (
+	"TUM-Live/tools"
+	"errors"
+	"fmt"
+	"github.com/getsentry/sentry-go"
+	"github.com/gin-gonic/gin"
+	"net/http"
+)
+
+func configGinStreamRestRouter(router *gin.Engine) {
+	g := router.Group("/")
+	g.Use(tools.InitStream)
+	g.Use(tools.AdminOfCourse)
+	g.GET("/api/stream/:streamID", getStream)
+}
+
+func getStream(c *gin.Context) {
+	foundContext, exists := c.Get("TUMLiveContext")
+	if !exists {
+		sentry.CaptureException(errors.New("context should exist but doesn't"))
+		c.AbortWithStatus(http.StatusInternalServerError)
+		return
+	}
+	tumLiveContext := foundContext.(tools.TUMLiveContext)
+	stream := *tumLiveContext.Stream
+	course := *tumLiveContext.Course
+	c.JSON(http.StatusOK,
+		gin.H{"course": course.Name,
+			"courseID":    course.ID,
+			"streamID":    stream.ID,
+			"name":        stream.Name,
+			"description": stream.Description,
+			"start":       stream.Start,
+			"end":         stream.End,
+			"ingest":      fmt.Sprintf("%sstream?secret=%s", tools.Cfg.IngestBase, stream.StreamKey),
+			"live":        stream.LiveNow,
+			"vod":         stream.Recording})
+}
